Extract code to get the required version of shared
Because I will need it also in a second place.
